Prepare for projectile object management

+++ Assets/Scripts/Object/Aircrafts/HitPoint.cs
@@ -3,20 +3,20 @@ using UnityEngine;
 
 public class HitPoint
 {
-    public int MaxHP { get; }
-    public int CurrentHP { get; private set; }
+    public float MaxHP { get; }
+    public float CurrentHP { get; private set; }
 
     public event Action OnDead;
 
-    public HitPoint(int maxHP, int? currentHP = null)
+    public HitPoint(float maxHP, float? currentHP = null)
     {
         MaxHP = maxHP;
         CurrentHP = currentHP.HasValue ? currentHP.Value : maxHP;
     }
 
-    public void IncreaseHP(int amount) => CurrentHP = Math.Clamp(CurrentHP + amount, 0, MaxHP);
+    public void IncreaseHP(float amount) => CurrentHP = Math.Clamp(CurrentHP + amount, 0, MaxHP);
 
-    public void DecreaseHP(int amount)
+    public void DecreaseHP(float amount)
     {
         CurrentHP = Math.Clamp(CurrentHP - amount, 0, CurrentHP);
